+++ anaconda_linter/lint/check_build_help.py
@@ -105,11 +105,9 @@ class uses_setuptools(LintCheck):
 
     """
 
-    severity = WARNING
-
     def check_recipe(self, recipe):
         if "setuptools" in recipe.get_deps("run"):
-            self.message()
+            self.message(severity=WARNING)
 
 
 class missing_wheel(LintCheck):
@@ -225,12 +223,10 @@ class avoid_noarch(LintCheck):
 
     """
 
-    severity = WARNING
-
     def check_recipe(self, recipe):
         noarch = recipe.get("build/noarch", "")
         if noarch == "python":
-            self.message(section="build")
+            self.message(section="build", severity=WARNING)
 
 
 class patch_unnecessary(LintCheck):
@@ -361,8 +357,6 @@ class remove_python_pinning(LintCheck):
 class gui_app(LintCheck):
     """This may be a GUI application. It is advised to test the GUI."""
 
-    severity = INFO
-
     guis = (
         "enaml",
         "glue-core",
@@ -380,4 +374,4 @@ class gui_app(LintCheck):
 
     def check_recipe(self, recipe):
         if set(self.guis).intersection(set(recipe.get_deps("run"))):
-            self.message()
+            self.message(severity=INFO)